industrial-rag/requirements.txt
langchain==0.1.0
weaviate-client==3.24.1
openai==1.3.0
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
tiktoken==0.5.1
numpy==1.24.3
pandas==2.0.3
torch==2.1.0
transformers==4.34.1
sentence-transformers==2.2.2